+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Removed support for ruby 1.8.x
 * Coercions#subdomain? and Coercions#belongs_to? are now protected
 * Myrrha::Domain is renamed as Myrrha::Domain::Native
+* Myrrha.domain has been removed. Use Myrrha::Domain.native instead.
 * In case of coercion failure, Myrrha::Error keeps the first coercion error under `cause`
   (that might be nil if no rule was triggered or no rule explcitely failed).